+++ CHANGELOG.md
@@ -17,6 +17,7 @@
     - `stylelint-default-less.json`
     - `stylelint-default-scss.json`
 - Added new opt `--upgrade` to `odoo-helper install py-tools` and `odoo-helper install dev-tools` commands
+- Added new opt `--update` to `odoo-helper install js-tools`
 - Added new opts to `odoo-helper db create` command:
     - `--tdb` - create test database with auto-generated name
     - `--name <name>` - allows to specify name of database as option